Split ChatSettingsAction from ChatSettings to avoid cyclic dependency (a dependency of ChatSettingsAction has a dependency on ChatSettings)
